+++ extensions/uptime-robot/CHANGELOG.md
@@ -1,5 +1,9 @@
 # UptimeRobot Changelog
 
+## [Delete Monitor] - {PR_MERGE_DATE}
+
+- Delete Monitor
+
 ## [Initial Version] - 2025-02-10
 
 - View Monitors